+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- Add support for C++ module files.
 - Add support for `.manifest` files.
 - Add support for MASM assembler files.
 - Add support for `.natvis` files.